Add movie titles to DOM and change listStyle height to fit list size
